+++ testing/testHeader.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "../src/City.cpp"
+#include "../src/Population.cpp"
+#include "../src/Route.cpp"
+#include <assert.h>
+
+void testRute(void);
+void testPopulation(Population* p);
+void testCity(void);
